Don't use List<T>.Find
git-svn-id: http://protobuf-net.googlecode.com/svn/trunk@41 b72047a4-3652-0410-9617-b3994939e97b

+++ protobuf-net/Base128Variant.cs
@@ -6,8 +6,8 @@ namespace ProtoBuf
 {
     internal static class Base128Variant
     {
-        internal const long  INT64_MSB = 1 << 63;
-        internal const int   INT32_MSB = 1 << 31;
+        internal const long  INT64_MSB = ((long)1) << 63;
+        internal const int   INT32_MSB = ((int)1) << 31;
 
         public static int EncodeInt32(int value, SerializationContext context)
         {
